// client-cargo-main (1)/client-cargo-main/src/stores/Trip/Trip.interface.ts
export interface ITripStore {
  classCosts: ITripTariff[];
  externalPrices: TExternalPrices[];
  personalCarsCosts: ITripTariff[];
  selfEmployee?: EmployeeModel;
  currentTripRequest?: TripRequestModel | null;
  currentTripRequestId?: string | null;
  tripRequestList: TripRequestModel[];
  pagination: TripPaginationModel;
  savedSuccessfully: boolean;
  purposes: TripPurpose[];
  tariffsTaxi: TTariffTaxi[];
  tariffsPersonal: TTariffPersonal[];
  tariffsPublic?: TTariffPublic;
  purposesMapped: Dictionary<TripPurpose>;
  delegatesByEmployeeId: Record<string, DelegateModel[]>;
  coopTripsSettings: TCoopTripsSettings[];
  suitableCooperativeTrips: TripSuitableModel[];
  nonTerminalTotalElements: number | null;
  isSearchingSuitableTrip: boolean;
  isSearchedSuitableTrip: boolean;
  actualTariff?: TTariffPublic;
  tripStatusChanged?: boolean;

  getCosts(data: ITripCalculateRequest): Promise<void>;
  getPersonalCarCost(data: ITripCalculateRequest): Promise<ITripTariff[] | undefined>;
  getTripRequestById(id: string): TripRequestModel | undefined;
  getExternalPrices(data: TTripRequestNew): Promise<void>;

  loadRequestList(): Promise<void>;
  loadRequestListTerminal(data: ITripRequestData): Promise<void>;
  loadRequestListNonTerminal(data: ITripRequestData): Promise<void>;
  loadPurposesList(orgId: string): Promise<void>;
  loadCoopTripsSettings(organizationId: string): Promise<void>;
  loadActualTariff(transTypeId: string, tariffId: UUID): Promise<void>;

  saveTripRequest(data: TTripRequestNew): Promise<void>;
  saveFile(file: File, folder: UUID | string): Promise<SavedFileInfo>;
  saveConfirmSuburbTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveConfirmCardTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveTravelCardRequest(data: InnerCityTransportRequestInfo): Promise<number>;
  saveSuburbCompensation(data: InnerCityTransportRequestInfo): Promise<number>;
  saveCityTripCompensation(data: CityTripCompensation): Promise<number>;
  savePublicTripRequest(data: PublicTripCompensation): Promise<number>;

  clearCoopTrips(): void;
  clearRequestList(): void;
  clearCurrentRequest(): void;
  clearClassCosts(): void;
  clearPersonalCarsCosts(): void;

  initStore(): void;

  //

  rateTripRequest(data: IRequestRating, requestId: string): Promise<void>;
  setCurrentTripRequest(reqId: string): void;

  searchCoopTrips(requestTrip: TTripRequestNew): Promise<void>;

  updateTripStateById({ id, newState }: { id: string; newState: Partial<TripRequestModel> }): void;

  joinCoopTrip(sharedId: string, data: TTripRequestNew): Promise<void>;
  addPersonalCost(tariff: ITripTariff): void;

  finishTripRequest(id: string): Promise<void>;
}
export interface ITripRequestData {
  pageSetting: { page: number; size: number };
  desiredDate?: { start?: number | null; end?: number | null };
  creationDate?: { start?: number | null; end?: number | null };
  page?: { sort: { sorted: boolean } };
  sortSetting?: { directionAsc: boolean };
  transportTypeEnum?: TransportTypeEnum;
}

export interface ITripService {
  calculateTripCost: (
    data: ITripCalculateRequest,
    callBack: (transportTypeId: string, tariffId: UUID) => {}
  ) => Promise<ITripTariff[]>;
  calculateCostForPersonalCarTrip: (data: ITripCalculateRequest) => Promise<ITripTariff[]>;
  calculateExternalPrices: (data: TTripRequestNew) => Promise<TExternalPrices[]>;

  saveTripRequest(data: TTripRequestNew): Promise<TripRequest | number>;

  editTripRequest(data: TTripRequestNew, requestId: string): Promise<number>;
  rateTripRequest(data: IRequestRating, requestId: string): Promise<number>;
  getTripRequestList(passengerId: string): Promise<TripRequest[]>;
  getTripRequestListNonTerminal(data: ITripRequestData): Promise<TripPaginationModel & { content: TripRequest[] }>;
  getTripRequestListTerminal(data: ITripRequestData): Promise<TripPaginationModel & { content: TripRequest[] }>;
  getTripRequest(requestId: string): Promise<TripRequest>;

  getAllPurposesByEmployee(orgId: string): Promise<TripPurpose[]>;
  getAllTariffsTaxi(): Promise<TTariffTaxi[]>;
  getAllTariffsPersonal(): Promise<TTariffPersonal[]>;
  getAllTariffsPublic(): Promise<TTariffPublic>;
  getTariffById(transTypeId: string, tariffId: UUID): Promise<TTariffPublic>;

  getAllCoopTrips(data: TripRequest): Promise<TTripCoop[]>;
  joinCoopTrip(sharedId: string, data: TTripRequestNew): Promise<TripRequest>;

  finishTripRequest(reqId: string): Promise<number>;

  saveFile(file: any, folder: UUID | string): Promise<SavedFileInfo>;
  saveConfirmSuburbTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveConfirmCardTripFile(requestId: string | UUID, filesData: SavedFileInfo[]): Promise<number>;
  saveTravelCardRequest(data: InnerCityTransportRequestInfo): Promise<number>;
  saveSuburbCompensation(data: InnerCityTransportRequestInfo): Promise<number>;
  saveCityTripCompensation(data: CityTripCompensation): Promise<number>;
  savePublicTripRequest(data: PublicTripCompensation): Promise<number>;
  loadCoopTripsSettings(organizationId: string): Promise<TCoopTripsSettings[]>;
}

export type ITripServiceResponse = Record<string, any>;

export interface ITripCalculateRequest {
  distance: number;
  time: number;
  tripDate: number;
  startPoint: Record<string, any>;
  organizationId: string;
  engineVolume?: number;
  waitingTime: number;
  intermediateWaitingTime: number;
}

export type Transport = TaxiClassEnum | TransportTypeEnum;

export enum TaxiClassEnum {
  ECONOMY = 'ECONOMY',
  COMFORT = 'COMFORT',
  PERSONAL = 'PERSONAL',
  BUSINESS = 'BUSINESS',
  TAXI = 'TAXI',
  CARSHARING = 'CARSHARING',
  BICYCLE = 'BICYCLE',
  WALK = 'WALK',
  PUBLIC = 'PUBLIC',
  SCOOTER = 'SCOOTER',
  YANDEX = 'YANDEX',
  CITYMOBIL = 'CITYMOBIL',
  UBER = 'UBER',

  VIP_BUS = 'VIP_BUS',
  SMALL_BUS = 'SMALL_BUS',
  MIDDLE_BUS = 'MIDDLE_BUS',
  LARGE_BUS = 'LARGE_BUS',
}
// ...
export type TTaxiClass = keyof typeof TaxiClassEnum;
// ...
export type TExternalPrices = t.TypeOf<typeof IOExternalPrices>;

export interface ITaxi {
  name: string;
  transportType: TransportTypeEnum;
  taxiClass?: TTaxiClass;
  waitingTime: number;
  disabled?: boolean;
  info?: string;
  type?: LIMIT_TYPE;
  contractorId?: string;
}

export interface ITripTariff {
  id: string;
  cost: number;
  transportType: { id: string; name: TransportTypeEnum };
  taxiClass?: TTaxiClass;
  contractorId?: string;

  bonusCost?: number | undefined;
}
// ...
export const IOTariffTaxi = t.intersection([
  IOTariffBase,
  t.type({
    taxiClass: ioTypeFromEnum<TaxiClassEnum>('TaxiClassEnum', TaxiClassEnum),
    rideCostPerMin: t.number,
    waitCostPerMin: t.number,
    freeWaitingTime: t.number,
  }),
  t.partial({
    id: t.string,
    priceDetails: t.unknown,
    deptStruct: t.string,
    deptName: t.string,
    vehicleType: t.string,
    minRideCost: t.number,
    savingsDeviationPct: t.number,
    distanceDeviationKm: t.number,
    timeDeviationMin: t.number,
    maxCapacity: t.number,
    minCancelTimeMin: t.number,
    tollRoads: t.boolean,
    historicalTraffic: t.boolean,
    minServiceTime: t.number,
    tariffId: t.string,
    active: t.boolean,
    contractorId: t.string,
  }),
]);
// ...
export type TTariffTaxi = t.TypeOf<typeof IOTariffTaxi>;

export type TTariffPersonal = t.TypeOf<typeof IOTariffPersonal>;
// ...
export type TTariffPublic = t.TypeOf<typeof IOTariffPublic>;

// TODO: Change to UUID

export const IOTripPurposeAttribute = t.intersection([
  t.type({
    name: t.string,
  }),
  t.partial({
    status: IOEmployeeStatusType,
    id: t.string,
  }),
]);

export const IOTripPurposeAttributes = t.type({
  attribute: IOTripPurposeAttribute,
});

export const IOTripPurposeDates = t.intersection([
  t.strict({
    // id: tt.uuid,
    id: t.string,
  }),
  t.partial({
    startDate: t.string,
    endDate: t.string,
  }),
]);

export const IOTripPurposeTimes = t.intersection([
  t.strict({
    // id: tt.uuid,
    id: t.string,
  }),
  t.partial({
    startTime: t.string,
    endTime: t.string,
  }),
]);
// ...
export const IOTripPurposeWeekdays = t.intersection([
  t.strict({
    // id: tt.uuid,
    id: t.string,
  }),
  t.partial({
    weekday: IOWeekdaysType,
  }),
]);

export const TripPurpose = t.intersection([
  t.type({
    // id: tt.uuid,
    id: t.string,
    label: t.string,
  }),
  t.partial({
    tripPurposeAttributes: t.array(IOTripPurposeAttributes),
    tripPurposeDates: t.array(IOTripPurposeDates),
    tripPurposeTimes: t.array(IOTripPurposeTimes),
    tripPurposeWeekdays: t.array(IOTripPurposeWeekdays),
  }),
]);

export type TripPurpose = t.TypeOf<typeof TripPurpose>;

export const IODeclineReason = t.type({
  reason: t.string,
  code: t.number,
});

export type IDeclineReason = t.TypeOf<typeof IODeclineReason>;

// Ratings

export interface IRequestRating {
  rating?: number;
  advantages?: string[];
  drawbacks?: string[];
  ratingComment?: string;
}

export const IORequestRating = t.partial({
  rating: t.number,
  advantages: t.array(t.string),
  drawbacks: t.array(t.string),
  ratingComment: t.string,
});

export type TRequestRating = t.TypeOf<typeof IORequestRating>;

// Trip Requests

export const TripRequest = t.intersection([
  IOHumanReadable,
  t.type({
    id: t.string,
    author: Employee,
    passenger: Employee,
    purpose: TripPurpose,
    passengerCount: t.number,
    desiredDate: t.number,
    creationTime: t.union([t.string, t.number]),
    approvedBy: Employee,
    coopTrip: t.boolean,
  }),
  t.partial({
    commentForDriver: t.string,
    humanReadableId: t.string,
    restOfLimit: t.number,
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
    status: TTripRequestStatuses,
    approvalState: ApprovalStateStatuses,
    expected: RequestRoute,
    taxiClass: ioTypeFromEnum<TaxiClassEnum>('TaxiClassEnum', TaxiClassEnum),
    tariffId: t.string,
    requestRating: IORequestRating,
    requestOptions: t.array(t.string),
    passengerLimitRemainder: t.array(LimitSharing),
    magentaOrderId: t.number,
    sharedRideId: t.number,
    occupiedPlacesCount: t.number,
    personalCar: IOPersonalCar,
    compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
    timeZone: t.string,
  }),
]);

export type TripRequest = t.TypeOf<typeof TripRequest>;

export type TTripRequestNew = Pick<TripRequest, 'author' | 'passenger' | 'transportType' | 'expected' | 'purpose'> &
  Partial<TripRequest>;
// ...
export type TTripCoop = t.TypeOf<typeof IOTripCoop>;
// ...
export const SavedFileInfo = t.strict({
  fileFormat: ioTypeFromEnum<UsedFileFormatEnum>('UsedFileFormatEnum', UsedFileFormatEnum),
  fileName: t.string,
  fileSize: t.number,
  folder: tt.uuid,
  id: tt.uuid,
});

export type SavedFileInfo = t.TypeOf<typeof SavedFileInfo>;

export const InnerCityTransportRequestInfo = t.strict({
  author: Employee,
  passenger: Employee,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  desiredDate: t.number,
  expected: RequestRoute,
  segments: t.array(Segment),
  waypoints: t.array(IOWaypoint),
  purpose: TripPurpose,
  compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
  compensationDocuments: t.array(SavedFileInfo),
});

export type InnerCityTransportRequestInfo = t.TypeOf<typeof InnerCityTransportRequestInfo>;

const TripData = t.strict({
  metroTicketsQuantity: t.number,
  tramTicketsQuantity: t.number,
  trolleybusTicketsQuantity: t.number,
  busTicketsQuantity: t.number,
});

export type TripData = t.TypeOf<typeof TripData>;

const tariffData = t.strict({
  metroTicketCost: t.number,
  tramTicketCost: t.number,
  trolleybusTicketCost: t.number,
  busTicketCost: t.number,
  metroAvailability: t.boolean,
  tramAvailability: t.boolean,
  trolleybusAvailability: t.boolean,
  busAvailability: t.boolean,
});

export const CityTripCompensation = t.strict({
  author: Employee,
  passenger: Employee,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  desiredDate: t.number,
  expected: RequestRoute,
  segments: t.array(Segment),
  waypoints: t.array(IOWaypoint),
  purpose: TripPurpose,
  compensationType: ioTypeFromEnum<TransportCompensations>('TransportCompensations', TransportCompensations),
  compensationDocuments: t.array(SavedFileInfo),

  tariffId: t.string,
  tripData: TripData,
  tariffData,
});

export type CityTripCompensation = t.TypeOf<typeof CityTripCompensation>;
// ...
export const PublicTripCompensation = t.intersection([
  t.type({
    author: Employee,
    passenger: Employee,
    purpose: TripPurpose,
    transportCompensation: t.array(TransportCompensation),
  }),
  t.partial({
    expected: RequestRoute,
    segments: t.array(Segment),
    waypoints: t.array(IOWaypoint),
    transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
    desiredDate: t.number,
    compensationDocuments: t.array(SavedFileInfo),
    tariffId: t.string,
    timeZone: t.string,
  }),
]);

export type PublicTripCompensation = t.TypeOf<typeof PublicTripCompensation>;

export const IOCoopTripsSettings = t.type({
  id: t.string,
  transportType: ioTypeFromEnum<TransportTypeEnum>('TransportTypeEnum', TransportTypeEnum),
  organizationId: t.string,
  economyIndicationYellowRangeLowerBorder: t.number,
  economyIndicationYellowRangeUpperBorder: t.number,
  settings: t.unknown,
});

export type TCoopTripsSettings = t.TypeOf<typeof IOCoopTripsSettings>;